+++ tiled/catalog/adapter.py
@@ -35,6 +35,7 @@ from sqlalchemy.ext.asyncio import create_async_engine
 from sqlalchemy.orm import selectinload
 from sqlalchemy.pool import AsyncAdaptedQueuePool
 from sqlalchemy.sql.expression import cast
+from sqlalchemy.sql.sqltypes import MatchType
 from starlette.status import HTTP_404_NOT_FOUND, HTTP_415_UNSUPPORTED_MEDIA_TYPE
 
 from tiled.queries import (
@@ -371,12 +372,24 @@ class CatalogNodeAdapter:
             return self.data_sources[0].structure
         return None
 
+    def apply_conditions(self, statement):
+        # IF this is a sqlite database and we are doing a full text MATCH
+        # query, we need a JOIN with the FTS5 virtual table.
+        if (self.context.engine.dialect.name == "sqlite") and any(
+            isinstance(condition.type, MatchType) for condition in self.conditions
+        ):
+            statement = statement.join(
+                orm.metadata_fts5, orm.metadata_fts5.c.rowid == orm.Node.id
+            )
+        for condition in self.conditions:
+            statement = statement.filter(condition)
+        return statement
+
     async def async_len(self):
         statement = select(func.count(orm.Node.key)).filter(
             orm.Node.ancestors == self.segments
         )
-        for condition in self.conditions:
-            statement = statement.filter(condition)
+        statement = self.apply_conditions(statement)
         async with self.context.session() as db:
             return (await db.execute(statement)).scalar_one()
 
@@ -398,17 +411,13 @@ class CatalogNodeAdapter:
             # Search queries and access controls apply only at the top level.
             assert not first_level.conditions
             return await first_level.lookup_adapter(segments[1:])
-        statement = (
-            select(orm.Node)
-            .filter(orm.Node.ancestors == self.segments + ancestors)
-            .options(
-                selectinload(orm.Node.data_sources).selectinload(
-                    orm.DataSource.structure
-                )
-            )
+        statement = select(orm.Node)
+        statement = self.apply_conditions(statement)
+        statement = statement.filter(
+            orm.Node.ancestors == self.segments + ancestors
+        ).options(
+            selectinload(orm.Node.data_sources).selectinload(orm.DataSource.structure)
         )
-        for condition in self.conditions:
-            statement = statement.filter(condition)
         async with self.context.session() as db:
             node = (await db.execute(statement.filter(orm.Node.key == key))).scalar()
         if node is None:
@@ -953,8 +962,7 @@ class CatalogContainerAdapter(CatalogNodeAdapter):
                 (offset + limit) if limit is not None else None,  # noqa: E203
             )
         statement = select(orm.Node.key).filter(orm.Node.ancestors == self.segments)
-        for condition in self.conditions:
-            statement = statement.filter(condition)
+        statement = self.apply_conditions(statement)
         async with self.context.session() as db:
             return (
                 (
@@ -976,8 +984,7 @@ class CatalogContainerAdapter(CatalogNodeAdapter):
                 (offset + limit) if limit is not None else None,  # noqa: E203
             )
         statement = select(orm.Node).filter(orm.Node.ancestors == self.segments)
-        for condition in self.conditions:
-            statement = statement.filter(condition)
+        statement = self.apply_conditions(statement)
         async with self.context.session() as db:
             nodes = (
                 (
@@ -1187,7 +1194,7 @@ def contains(query, tree):
 def full_text(query, tree):
     dialect_name = tree.engine.url.get_dialect().name
     if dialect_name == "sqlite":
-        raise UnsupportedQueryType("full_text")
+        condition = orm.metadata_fts5.c.metadata.match(query.text)
     elif dialect_name == "postgresql":
         tsvector = func.jsonb_to_tsvector(
             cast("simple", REGCONFIG), orm.Node.metadata_, cast(["string"], JSONB)