+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Write the date in place of the "Unreleased" in the case a new version is release
 ### Fixed
 
 - Updated `BaseClient.formats` to use the `dict` structure for specs.
+- The `tiled serve directory --watch` function was not compatible with recent `anyio`
 
 ## v0.1.0a122 (23 May 2024)