+++ CHANGELOG.md
@@ -3,6 +3,13 @@ Write the date in place of the "Unreleased" in the case a new version is release
 
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- Updated `BaseClient.formats` to use the `dict` structure for specs.
+- The `tiled serve directory --watch` function was not compatible with recent `anyio`
+
 ## v0.1.0a122 (23 May 2024)
 
 ### Fixed